moved useOrderSummary to its own file
