+++ packages/checkout/src/hooks/useCheckoutUI/useOrderSummary.tsx
@@ -4,30 +4,19 @@ import { formatDisplay, NetworkBadge } from '@0xsequence/connect'
 import { useGetContractInfo, useGetTokenMetadata, useGetCoinPrices } from '@0xsequence/hooks'
 import { ReactNode } from 'react'
 import { formatUnits } from 'viem'
-import { Collectible } from '../contexts/SelectPaymentModal'
+import { Collectible } from '../../contexts/SelectPaymentModal'
 
-// summary
-
-// crypto payment
-
-// credit card payment
-
-interface UseCheckoutUIArgs {
+export interface UseOrderSummaryArgs {
   chain: string | number
   currencyAddress: string
   totalPriceRaw: string
   collectibles: Collectible[]
   collectionAddress: string
-  recipientAddress: string
-  transactionConfirmations?: number
-  onSuccess?: (txHash: string) => void
-  onError?: (error: Error) => void
-  onClose?: () => void
 }
 
-interface UseOrderSummaryArgs {}
+export interface UseOrderSummaryConfig {}
 
-interface CollectibleItem {
+export interface CollectibleItem {
   quantityRaw: string
   quantityFormatted: string
   collectionName: string
@@ -35,7 +24,7 @@ interface CollectibleItem {
   collectibleImageUrl: string
 }
 
-interface UseOrderSummaryData {
+export interface UseOrderSummaryData {
   formattedCryptoPrice: string
   cryptoSymbol: string
   totalPriceFiat: string
@@ -45,38 +34,24 @@ interface UseOrderSummaryData {
   collectibleItems: CollectibleItem[]
 }
 
-interface UseOrderSummaryReturn {
+export interface UseOrderSummaryReturn {
   error: Error | null
   data: UseOrderSummaryData | null
   isLoading: boolean
 }
 
-interface UseCheckoutUIReturn {
-  useOrderSummary: (args: UseOrderSummaryArgs) => UseOrderSummaryReturn
-}
+export const useOrderSummary =
+  ({
+    chain,
+    currencyAddress,
+    totalPriceRaw,
+    collectibles,
+    collectionAddress
+  }: UseOrderSummaryArgs): (() => UseOrderSummaryReturn) =>
+  () => {
+    const network = findSupportedNetwork(chain)
+    const chainId = network?.chainId || 137
 
-export const useCheckoutUI = ({
-  chain,
-  currencyAddress,
-  totalPriceRaw,
-  collectibles,
-  collectionAddress,
-  recipientAddress,
-  transactionConfirmations
-}: UseCheckoutUIArgs): UseCheckoutUIReturn => {
-  const network = findSupportedNetwork(chain)
-  const chainId = network?.chainId || 137
-
-  const {
-    data: currencyInfo,
-    isLoading: isLoadingCurrencyInfo,
-    error: errorCurrencyInfo
-  } = useGetContractInfo({
-    chainID: String(chainId),
-    contractAddress: currencyAddress
-  })
-
-  const useOrderSummary = () => {
     const {
       data: tokenMetadatas,
       isLoading: isLoadingTokenMetadatas,
@@ -105,6 +80,15 @@ export const useCheckoutUI = ({
     } = useGetContractInfo({
       chainID: String(chainId),
       contractAddress: collectionAddress
+    })
+
+    const {
+      data: currencyInfo,
+      isLoading: isLoadingCurrencyInfo,
+      error: errorCurrencyInfo
+    } = useGetContractInfo({
+      chainID: String(chainId),
+      contractAddress: currencyAddress
     })
 
     const isLoading = isLoadingCurrencyInfo || isLoadingTokenMetadatas || isLoadingCoinPrices || isLoadingCollectionInfo
@@ -146,6 +130,3 @@ export const useCheckoutUI = ({
 
     return { isLoading, error, data }
   }
-
-  return { useOrderSummary }
-}
